requirements.txt
# Streamlit App
streamlit
pandas
numpy
openpyxl
Pillow

# PDF Processing
pdfplumber
PyMuPDF
fitz  # alias for PyMuPDF (optional install hint)

# OCR & Layout
easyocr
layoutparser[layoutmodels,tesseract]  # uses paddleocr internally
paddleocr
opencv-python-headless

# LangChain & Ollama
langchain
langchain-core
langchain-community
langchain-text-splitters
langchainhub

# Ollama embeddings & LLM
faiss-cpu
nomic
sentence-transformers

# Others
tqdm
layoutparser[ocr]
paddleocr
paddlepaddle
